pyproject.toml
[tool.poetry]
name = "devkami_buzzmaker"
version = "0.1.0"
description = ""
authors = ["sweemeng <sweester@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
PyYAML = "^6.0"
requests = "^2.26.0"
beautifulsoup4 = "^4.10.0"
nltk = "^3.6.7"
lxml = "^4.7.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
